fixed some depreciation warnings

+++ optics/filters.py
@@ -9,7 +9,7 @@ def stockman(minLambda=390, maxLambda=770, ONLY_LENS=False,
             ONLY_MACULA=False, RETURN_SPECTRUM=False, resolution=1):
     '''
     '''
-    dsamp = resolution * 10.0
+    dsamp = int(resolution * 10.0)
     lens = np.genfromtxt(this_dir + '/stockman/lens.csv', 
                          delimiter=',')[::dsamp, :]
     macula = np.genfromtxt(this_dir+ '/stockman/macular.csv', 
@@ -17,8 +17,8 @@ def stockman(minLambda=390, maxLambda=770, ONLY_LENS=False,
 
     spectrum = lens[:, 0]
     # find location of maxLambda
-    ind1 = np.where(spectrum == minLambda)[0]
-    ind2 = np.where(spectrum == maxLambda)[0]
+    ind1 = int(np.where(spectrum == minLambda)[0])
+    ind2 = int(np.where(spectrum == maxLambda)[0])
 
     spectrum = spectrum[ind1:ind2 + 1]
     filters = (10.0 ** (lens[ind1:ind2 + 1, 1] +
